+++ lib/formix.dart
@@ -44,4 +44,5 @@ export 'src/widgets/field_selector/formix_field_value_selector.dart';
 export 'src/validators/validators.dart';
 export 'src/analytics/form_analytics.dart';
 export 'src/analytics/logging_form_analytics.dart';
+export 'src/widgets/dependent_async_field.dart';
 export 'src/widgets/form_registry.dart';